faketrello: Improved the dragging animation

+++ examples/faketrello/views/Card.qml
@@ -4,18 +4,18 @@ import "../components"
 Item {
     id: card
 
+    state: "Default"
+
     property int normalHeight: 48
-    height: (dropArea.containsDrag && !dropArea.sameList) ? normalHeight * 2 : normalHeight
+    height: normalHeight
 
     property string listUuid
     property string cardUuid
     property string text
+    property int cardIndex: -1
 
-    Behavior on height {
-        NumberAnimation {
-            duration: 200
-            easing.type: Easing.OutQuad;
-        }
+    function inSameList(c1, c2) {
+        return c1.ListView.view === c2.ListView.view;
     }
 
     MouseArea {
@@ -24,41 +24,53 @@ Item {
         hoverEnabled: true
         drag.target: container
         onReleased: {
-            container.Drag.drop();
+            if (card.state === "Dragging") {
+                container.Drag.drop();
+                card.state = "Default";
+            }
         }
     }
 
     DropArea {
         id: dropArea
         anchors.fill: parent
-        property bool sameList: true
+        enabled: !container.Drag.active
         onEntered: {
+            if (card === drag.source) {
+                return;
+            }
 
-            sameList = (card.listUuid  === drag.source.listUuid);
             drag.accept(Qt.MoveAction);
+            drag.source.state = "Dragging";
 
-            if (card.listUuid !== drag.source.listUuid) {
+            if (inSameList(card, drag.source) && card.cardIndex >= drag.source.cardIndex) {
+                card.state = "DroppingAfter";
+            } else {
+                card.state = "DroppingBefore";
+            }
+        }
+
+        onExited: {
+            if (card === drag.source) {
                 return;
             }
 
-            if (card.cardUuid === drag.source.cardUuid) {
+            card.state = "Default";
+        }
+
+        onDropped: {
+            if (card === drag.source) {
                 return;
             }
+
+            card.state = "Dropped"
+            card.state = "Default";
+            drag.source.state = "Default";
 
             App.moveCard(drag.source.listUuid,
                          drag.source.cardUuid,
                          card.listUuid,
                          card.cardUuid);
-        }
-
-        onDropped: {
-            if (!sameList) {
-                drop.accept(Qt.MoveAction);
-                App.moveCard(drag.source.listUuid,
-                             drag.source.cardUuid,
-                             card.listUuid,
-                             card.cardUuid);
-            }
         }
     }
 
@@ -69,6 +81,7 @@ Item {
         anchors.horizontalCenter: parent.horizontalCenter
         anchors.bottom: parent.bottom
         anchors.bottomMargin: 8
+        anchors.topMargin: 8
 
         width: card.width - 16
         height: normalHeight - 16
@@ -115,7 +128,6 @@ Item {
                     easing.type: Easing.OutQuad;
                 }
             }
-
         }
 
         states: State {
@@ -137,7 +149,76 @@ Item {
                 anchors.bottom: undefined
             }
         }
-
     }
+
+    states: [
+        State {
+            name: "Default"
+        },
+        State {
+            name: "DroppingBefore"
+
+            PropertyChanges {
+                target: card
+                height: normalHeight * 2
+            }
+        },
+        State {
+            name: "DroppingAfter"
+
+            PropertyChanges {
+                target: card
+                height: normalHeight * 2
+            }
+
+            AnchorChanges {
+                target: container
+                anchors.bottom: undefined
+                anchors.top: card.top
+            }
+        },
+        State {
+            name: "Dropped"
+        },
+        State {
+            name: "Dragging"
+            PropertyChanges {
+                target: card
+                height: 0
+            }
+        }
+    ]
+
+    onStateChanged: {
+        console.log(card.text, card.state)
+    }
+
+    transitions: [
+        Transition {
+            from: "Default"
+            to: "DroppingAfter,DroppingBefore"
+            reversible: true
+
+            NumberAnimation {
+                target: card
+                property: "height"
+                duration: 200
+                easing.type: Easing.Linear
+                alwaysRunToEnd: true
+            }
+        },
+        Transition {
+            from: "Default"
+            to: "Dragging"
+            NumberAnimation {
+                target: card
+                property: "height"
+                duration: 200
+                easing.type: Easing.Linear
+                alwaysRunToEnd: true
+            }
+        }
+    ]
+
 }
